% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_print.R
\name{print.ssize_result}
\alias{print.ssize_result}
\title{Print method for ssize_result}
\usage{
\method{print}{ssize_result}(x, ...)
}
\arguments{
\item{x}{Object of class \code{ssize_result}}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for ssize_result
}
