% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_emperical.R
\name{sim_empirical_power}
\alias{sim_empirical_power}
\title{Simulate empirical power}
\usage{
sim_empirical_power(
  design = c("continuous", "binary", "survival"),
  nsim = 1000,
  n = 50,
  delta = NULL,
  sd = NULL,
  p1 = NULL,
  p2 = NULL,
  HR = NULL,
  lambdaC = NULL,
  accrual = NULL,
  followup = NULL,
  dropout = 0,
  alpha = 0.05,
  alternative = c("two.sided", "one.sided")
)
}
\arguments{
\item{design}{"continuous", "binary", or "survival".}

\item{nsim}{Number of simulations.}

\item{n}{Sample size per arm.}

\item{delta}{Mean difference for continuous outcome.}

\item{sd}{Standard deviation for continuous outcome.}

\item{p1, p2}{Probabilities for binary outcome.}

\item{HR}{Hazard ratio for survival.}

\item{lambdaC}{Baseline hazard in control group for survival.}

\item{accrual, followup, dropout}{For survival.}

\item{alpha}{Significance level.}

\item{alternative}{"two.sided" or "one.sided".}
}
\value{
Empirical power (proportion of significant tests).
}
\description{
Simulates parallel or survival trials to empirically estimate power under complex settings.
}
\examples{
sim_empirical_power("continuous", nsim=100, n=20, delta=1, sd=2)
}
